server/requirements.txt
Flask==3.0.3
PyYAML==6.0.1
requests==2.32.3